+++ sdk/src/main/java/com/bugsnag/android/DeliveryStyle.java
@@ -4,5 +4,6 @@ package com.bugsnag.android;
 enum DeliveryStyle {
     SAME_THREAD,
     ASYNC,
-    ASYNC_WITH_CACHE
+    ASYNC_WITH_CACHE,
+    NO_CACHE
 }